scale float

+++ src/ofxDmtrUI3.h
@@ -322,7 +322,7 @@ public:
 	bool hasPresets = false;
 
 	// talvez desencanar disso...
-	int scale = 1;
+	float scale = 1;
 	
 	
 	string loadedXmlFile = "";
